Refactor injecting functions to util
