+++ static/app/views/dashboards/widgetBuilder/buildSteps/columnsStep/index.tsx
@@ -7,10 +7,8 @@ import useCustomMeasurements from 'sentry/utils/useCustomMeasurements';
 import {getDatasetConfig} from 'sentry/views/dashboards/datasetConfig/base';
 import type {DisplayType, WidgetQuery, WidgetType} from 'sentry/views/dashboards/types';
 import {hasDatasetSelector} from 'sentry/views/dashboards/utils';
-import {appendFieldIfUnknown} from 'sentry/views/discover/table/queryField';
-import {FieldValueKind} from 'sentry/views/discover/table/types';
 
-import {DataSet, getFieldOptionFormat} from '../../utils';
+import {addIncompatibleFunctions, DataSet} from '../../utils';
 import {BuildStep} from '../buildStep';
 
 import {ColumnFields} from './columnFields';
@@ -42,7 +40,7 @@ export function ColumnsStep({
   const {customMeasurements} = useCustomMeasurements();
   const datasetConfig = getDatasetConfig(widgetType);
 
-  let fieldOptions = datasetConfig.getTableFieldOptions(
+  const fieldOptions = datasetConfig.getTableFieldOptions(
     organization,
     tags,
     customMeasurements
@@ -56,36 +54,7 @@ export function ColumnsStep({
     hasDatasetSelector(organization) &&
     [DataSet.ERRORS, DataSet.TRANSACTIONS].includes(dataSet)
   ) {
-    explodedFields.forEach(field => {
-      // Inject functions that aren't compatible with the current dataset
-      if (field.kind === 'function') {
-        const functionName = field.alias || field.function[0];
-        if (!(`function:${functionName}` in fieldOptions)) {
-          const formattedField = getFieldOptionFormat(field);
-          if (formattedField) {
-            const [key, value] = formattedField;
-            fieldOptions[key] = value;
-
-            // If the function needs to be injected, inject the parameter as a tag
-            // as well if it isn't already an option
-            if (
-              field.function[1] &&
-              !fieldOptions[`field:${field.function[1]}`] &&
-              !fieldOptions[`tag:${field.function[1]}`]
-            ) {
-              fieldOptions = appendFieldIfUnknown(fieldOptions, {
-                kind: FieldValueKind.TAG,
-                meta: {
-                  dataType: 'string',
-                  name: field.function[1],
-                  unknown: true,
-                },
-              });
-            }
-          }
-        }
-      }
-    });
+    addIncompatibleFunctions(explodedFields, fieldOptions);
   }
 
   return (
